+++ packages/mars-build/src/compiler/script/script-h5.js
@@ -23,7 +23,27 @@ exports.postCompile = function (file) {
 
 exports.compileRouter = function (content, options) {
     const {config = {}, mars} = options;
-    const routes = config.pages || [];
+    const {
+        pages = [],
+        subPackages = []
+    } = config;
+    let subPages = [];
+    if (subPackages && subPackages.length > 0) {
+        subPackages.forEach(item => {
+            if (!item.pages) {
+                return;
+            }
+            if (item.pages.length > 0) {
+                item.pages.forEach(route => {
+                    subPages.push(`${item.root}/${route}`);
+                });
+            }
+            else {
+                subPages.push(`${item.root}/${item.pages[0]}`);
+            }
+        });
+    }
+    const routes = pages.concat(subPages);
     const mode = mars && mars.mode ? mars.mode : 'history';
     const routerRet = babel.transform(content.toString(), {
         plugins: [transformRouterPlugin({
@@ -70,15 +90,32 @@ exports.compileMain = function (content, options) {
         window = null,
         pagesInfo
     } = mainOptions;
+    const {
+        backgroundColor,
+        borderStyle,
+        color,
+        selectedColor,
+        list = []
+    } = tabBar;
+    const tabBarStyle = {
+        backgroundColor,
+        borderStyle,
+        color,
+        selectedColor
+    };
     const mainRet = babel.transform(content.toString(), {
         plugins: [transformMainPlugin({
-            routes: tabBar && tabBar.list || [],
+            routes: list,
+            tabBarStyle,
             window,
             mars: {
-                navigationBarHomeColor: mars && mars.navigationBarHomeColor || 'dark',
-                showNavigationBorder: mars ? !!mars.showNavigationBorder : true,
-                useTransition: mars ? !!mars.useTransition : true,
-                homePage: `/${tabBar && tabBar.list && tabBar.list.length > 0 ? tabBar.list[0].pagePath : pages[0]}`
+                navigationBarHomeColor: mars.navigationBarHomeColor === undefined
+                    ? 'dark'
+                    : mars.navigationBarHomeColor,
+                showNavigationBorder: !!mars.showNavigationBorder,
+                useTransition: mars.useTransition === undefined ? true : mars.useTransition,
+                homePage: `/${list.length > 0 ? list[0].pagePath : pages[0]}`,
+                supportPWA: !!mars.supportPWA
             },
             componentSet,
             pagesInfo