+++ packages/mars-build/src/h5/transform/plugins/transformMainPlugin.js
@@ -7,15 +7,15 @@
 
 module.exports = function getVisitor(options = {}) {
     return ({types: t}) => {
-        const {routes, componentSet, pagesInfo, window, mars} = options;
+        const {routes, componentSet, pagesInfo, window, mars, tabBarStyle} = options;
         return {
             visitor: {
                 // 处理main.js 里的 render函数参数
                 ArrowFunctionExpression(path, state) {
-                    let tabBars = [];
+                    let tabBarList = [];
                     // import component & routes
                     for (let r of routes) {
-                        tabBars.push(t.objectExpression([
+                        tabBarList.push(t.objectExpression([
                             t.objectProperty(
                                 t.identifier('pagePath'),
                                 t.stringLiteral(`/${r.pagePath}`)
@@ -42,10 +42,28 @@ module.exports = function getVisitor(options = {}) {
                             )
                         ]));
                     }
+                    let styleArr = [];
+                    Object.keys(tabBarStyle).forEach(key => {
+                        if (tabBarStyle[key]) {
+                            styleArr.push(t.objectProperty(
+                                t.identifier(key),
+                                t.stringLiteral(`${tabBarStyle[key]}`)
+                            ));
+                        }
+                    });
                     let propsAst = [];
                     propsAst.push(t.objectProperty(
                         t.identifier('tabBars'),
-                        t.arrayExpression(tabBars)
+                        t.objectExpression([
+                            t.objectProperty(
+                                t.identifier('list'),
+                                t.arrayExpression(tabBarList)
+                            ),
+                            t.objectProperty(
+                                t.identifier('style'),
+                                t.objectExpression(styleArr)
+                            )
+                        ])
                     ));
                     // 配置 全部变量 window
                     Object.keys(window).forEach(key => {
@@ -150,6 +168,14 @@ module.exports = function getVisitor(options = {}) {
                             // 插入 import basicComponents from './components.js';
                             insertImportDeclaration('basicComponents', './components.js');
                         }
+                        // 注册 service worker
+                        if (!!mars.supportPWA) {
+                            path.node.body.unshift(t.expressionStatement(t.callExpression(
+                                t.identifier('registerServiceWorker'),
+                                [t.stringLiteral('sw.js')]
+                            )));
+                            insertImportDeclaration('registerServiceWorker', './registerServiceWorker.js');
+                        }
                         // 插入 import Vue from 'vue';
                         insertImportDeclaration('Vue', 'vue');
                     }
@@ -157,4 +183,4 @@ module.exports = function getVisitor(options = {}) {
             }
         };
     };
-}
+};